Cleaning up functional and performance tests and removing old code.
