+++ src/System.Text.Primitives/System/Text/Parsing/InvariantSigned.cs
@@ -988,103 +988,6 @@ namespace System.Text
                 return true;
             }
 
-            public static bool TryParseInt32_OLD(ReadOnlySpan<byte> text, out int value, out int bytesConsumed)
-            {
-                if (text.Length < 1)
-                {
-                    bytesConsumed = 0;
-                    value = default(int);
-                    return false;
-                }
-
-                int indexOfFirstDigit = 0;
-                int sign = 1;
-                if (text[0] == '-')
-                {
-                    indexOfFirstDigit = 1;
-                    sign = -1;
-                }
-                else if (text[0] == '+')
-                {
-                    indexOfFirstDigit = 1;
-                }
-
-                if (indexOfFirstDigit >= text.Length)
-                {
-                    bytesConsumed = 0;
-                    value = default(int);
-                    return false;
-                }
-
-                int overflowLength = Int32OverflowLength + indexOfFirstDigit;
-
-                // Parse the first digit separately. If invalid here, we need to return false.
-                int firstDigit = text[indexOfFirstDigit] - 48; // '0'
-                if (firstDigit < 0 || firstDigit > 9)
-                {
-                    bytesConsumed = 0;
-                    value = default(int);
-                    return false;
-                }
-                int parsedValue = firstDigit;
-
-                if (text.Length < overflowLength)
-                {
-                    // Length is less than Int32OverflowLength; overflow is not possible
-                    for (int index = indexOfFirstDigit + 1; index < text.Length; index++)
-                    {
-                        int nextDigit = text[index] - 48; // '0'
-                        if (nextDigit < 0 || nextDigit > 9)
-                        {
-                            bytesConsumed = index;
-                            value = parsedValue * sign;
-                            return true;
-                        }
-                        parsedValue = parsedValue * 10 + nextDigit;
-                    }
-                }
-                else
-                {
-                    // Length is greater than Int32OverflowLength; overflow is only possible after Int32OverflowLength
-                    // digits. There may be no overflow after Int32OverflowLength if there are leading zeroes.
-                    for (int index = indexOfFirstDigit + 1; index < overflowLength - 1; index++)
-                    {
-                        int nextDigit = text[index] - 48; // '0'
-                        if (nextDigit < 0 || nextDigit > 9)
-                        {
-                            bytesConsumed = index;
-                            value = parsedValue * sign;
-                            return true;
-                        }
-                        parsedValue = parsedValue * 10 + nextDigit;
-                    }
-                    for (int index = overflowLength - 1; index < text.Length; index++)
-                    {
-                        int nextDigit = text[index] - 48; // '0'
-                        if (nextDigit < 0 || nextDigit > 9)
-                        {
-                            bytesConsumed = index;
-                            value = parsedValue * sign;
-                            return true;
-                        }
-                        // If parsedValue > (int.MaxValue / 10), any more appended digits will cause overflow.
-                        // if parsedValue == (int.MaxValue / 10), any nextDigit greater than 7 or 8 (depending on sign) implies overflow.
-                        bool positive = sign > 0;
-                        bool nextDigitTooLarge = nextDigit > 8 || (positive && nextDigit > 7);
-                        if (parsedValue > int.MaxValue / 10 || parsedValue == int.MaxValue / 10 && nextDigitTooLarge)
-                        {
-                            bytesConsumed = 0;
-                            value = 0;
-                            return false;
-                        }
-                        parsedValue = parsedValue * 10 + nextDigit;
-                    }
-                }
-
-                bytesConsumed = text.Length;
-                value = parsedValue * sign;
-                return true;
-            }
             #endregion
 
             #region Int64
