+++ src/System.Text.Primitives/System/Text/Parsing/Signed.cs
@@ -24,9 +24,9 @@ namespace System.Text
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static bool IsInvalid(uint i)
+        private static bool IsValid(uint i)
         {
-            return i > (uint)TextEncoder.Symbol.D9 || i < (uint)TextEncoder.Symbol.D0;
+            return i <= (uint)TextEncoder.Symbol.D9;
         }
 
         // If parsedValue > (sbyte.MaxValue / 10), any more appended digits will cause overflow.
@@ -181,7 +181,6 @@ namespace System.Text
             return true;
         }
 
-
         public static bool TryParseInt16(ReadOnlySpan<byte> text, out short value, out int bytesConsumed, TextFormat format = default(TextFormat), TextEncoder encoder = null)
         {
             encoder = encoder == null ? TextEncoder.Utf8 : encoder;
@@ -296,7 +295,6 @@ namespace System.Text
             return true;
         }
 
-
         public static bool TryParseInt32(ReadOnlySpan<byte> text, out int value, out int bytesConsumed, TextFormat format = default(TextFormat), TextEncoder encoder = null)
         {
             bool isDefault = format.IsDefault;
@@ -360,7 +358,7 @@ namespace System.Text
             }
 
             int answer = 0;
-            if (symbol <= (uint)TextEncoder.Symbol.D9)
+            if (IsValid(symbol))
             {
                 int numBytes = consumed;
                 if (symbol == (uint)TextEncoder.Symbol.D0)
@@ -371,7 +369,7 @@ namespace System.Text
                         if (index >= textLength) goto Done;
                         if (!encoder.TryParseSymbol(ref Unsafe.Add(ref textByte, index), out symbol, out consumed)) goto Done;
                     } while (symbol == (uint)TextEncoder.Symbol.D0);
-                    if (IsInvalid(symbol)) goto Done;
+                    if (!IsValid(symbol)) goto Done;
                 }
 
                 int firstNonZeroDigitIndex = index;
@@ -383,7 +381,7 @@ namespace System.Text
                         index += consumed;
                         if (index >= textLength) goto Done;
                         if (!encoder.TryParseSymbol(ref Unsafe.Add(ref textByte, index), out symbol, out consumed)) goto Done;
-                    } while (symbol <= (uint)TextEncoder.Symbol.D9);
+                    } while (IsValid(symbol));
                 }
                 else
                 {
@@ -394,7 +392,7 @@ namespace System.Text
                         if (index - firstNonZeroDigitIndex == Int32OverflowLength * numBytes - 1)
                         {
                             if (!encoder.TryParseSymbol(ref Unsafe.Add(ref textByte, index), out symbol, out consumed)) goto Done;
-                            if (symbol <= (uint)TextEncoder.Symbol.D9)
+                            if (IsValid(symbol))
                             {
                                 symbol -= (int)TextEncoder.Symbol.D0;
                                 if (WillOverFlow(answer, (int)symbol, sign)) goto FalseExit;
@@ -404,7 +402,7 @@ namespace System.Text
                             goto Done;
                         }
                         if (!encoder.TryParseSymbol(ref Unsafe.Add(ref textByte, index), out symbol, out consumed)) goto Done;
-                    } while (symbol <= (uint)TextEncoder.Symbol.D9);
+                    } while (IsValid(symbol));
                 }
                 goto Done;
             }
@@ -420,133 +418,6 @@ namespace System.Text
             return true;
         }
 
-        public static bool TryParseInt32_OLD(ReadOnlySpan<byte> text, out int value, out int bytesConsumed, TextFormat format = default(TextFormat), TextEncoder encoder = null)
-        {
-            encoder = encoder == null ? TextEncoder.Utf8 : encoder;
-
-            if (!format.IsDefault && format.HasPrecision)
-            {
-                throw new NotImplementedException("Format with precision not supported.");
-            }
-
-            if (encoder.IsInvariantUtf8)
-            {
-                if (format.IsHexadecimal)
-                {
-                    return InvariantUtf8.Hex.TryParseInt32(text, out value, out bytesConsumed);
-                }
-                else
-                {
-                    return InvariantUtf8.TryParseInt32(text, out value, out bytesConsumed);
-                }
-            }
-            else if (encoder.IsInvariantUtf16)
-            {
-                ReadOnlySpan<char> utf16Text = text.NonPortableCast<byte, char>();
-                int charsConsumed;
-                bool result;
-                if (format.IsHexadecimal)
-                {
-                    result = InvariantUtf16.Hex.TryParseInt32(utf16Text, out value, out charsConsumed);
-                }
-                else
-                {
-                    result = InvariantUtf16.TryParseInt32(utf16Text, out value, out charsConsumed);
-                }
-                bytesConsumed = charsConsumed * sizeof(char);
-                return result;
-            }
-
-            if (format.IsHexadecimal)
-            {
-                throw new NotImplementedException("The only supported encodings for hexadecimal parsing are InvariantUtf8 and InvariantUtf16.");
-            }
-
-            if (!(format.IsDefault || format.Symbol == 'G' || format.Symbol == 'g'))
-            {
-                throw new NotImplementedException(String.Format("Format '{0}' not supported.", format.Symbol));
-            }
-
-            if (text.Length < 1)
-            {
-                bytesConsumed = 0;
-                value = default(int);
-                return false;
-            }
-
-            uint nextSymbol;
-            int thisSymbolConsumed;
-            if (!encoder.TryParseSymbol(text, out nextSymbol, out thisSymbolConsumed))
-            {
-                value = default(int);
-                bytesConsumed = 0;
-                return false;
-            }
-
-            int sign = 1;
-            if ((TextEncoder.Symbol)nextSymbol == TextEncoder.Symbol.MinusSign)
-            {
-                sign = -1;
-            }
-
-            int signConsumed = 0;
-            if ((TextEncoder.Symbol)nextSymbol == TextEncoder.Symbol.PlusSign || (TextEncoder.Symbol)nextSymbol == TextEncoder.Symbol.MinusSign)
-            {
-                signConsumed = thisSymbolConsumed;
-                if (signConsumed >= text.Length)
-                {
-                    bytesConsumed = 0;
-                    value = default(int);
-                    return false;
-                }
-                if (!encoder.TryParseSymbol(text.Slice(signConsumed), out nextSymbol, out thisSymbolConsumed))
-                {
-                    value = default(int);
-                    bytesConsumed = 0;
-                    return false;
-                }
-            }
-
-            if (nextSymbol > 9)
-            {
-                value = default(int);
-                bytesConsumed = 0;
-                return false;
-            }
-
-            int parsedValue = (int)nextSymbol;
-            int index = signConsumed + thisSymbolConsumed;
-
-            while (index < text.Length)
-            {
-                bool success = encoder.TryParseSymbol(text.Slice(index), out nextSymbol, out thisSymbolConsumed);
-                if (!success || nextSymbol > 9)
-                {
-                    bytesConsumed = index;
-                    value = (int)(parsedValue * sign);
-                    return true;
-                }
-
-                // If parsedValue > (int.MaxValue / 10), any more appended digits will cause overflow.
-                // if parsedValue == (int.MaxValue / 10), any nextDigit greater than 7 or 8 (depending on sign) implies overflow.
-                bool positive = sign > 0;
-                bool nextDigitTooLarge = nextSymbol > 8 || (positive && nextSymbol > 7);
-                if (parsedValue > int.MaxValue / 10 || (parsedValue == int.MaxValue / 10 && nextDigitTooLarge))
-                {
-                    bytesConsumed = 0;
-                    value = default(int);
-                    return false;
-                }
-
-                index += thisSymbolConsumed;
-                parsedValue = parsedValue * 10 + (int)nextSymbol;
-            }
-
-            bytesConsumed = text.Length;
-            value = (int)(parsedValue * sign);
-            return true;
-        }
-        
         public static bool TryParseInt64(ReadOnlySpan<byte> text, out long value, out int bytesConsumed, TextFormat format = default(TextFormat), TextEncoder encoder = null)
         {
             encoder = encoder == null ? TextEncoder.Utf8 : encoder;
